+++ pyxem/data/_data.py
@@ -85,7 +85,7 @@ def pdnip_glass(allow_download=False, **kwargs):  # pragma: no cover
     return hs.load(store, **kwargs)
 
 
-def zrnb_percipitate(allow_download=False, **kwargs):  # pragma: no cover
+def zrnb_precipitate(allow_download=False, **kwargs):  # pragma: no cover
     """A small 4-D STEM dataset of a ZrNb precipitate for strain mapping.
 
     Data liscenced under CC BY 4.0
@@ -107,8 +107,8 @@ def zrnb_percipitate(allow_download=False, **kwargs):  # pragma: no cover
     """
     import zarr
 
-    pdnip = Dataset("PdNiP.zspy")
-    file_path = pdnip.fetch_file_path(allow_download=allow_download)
+    zrnb = Dataset("ZrNbPercipitate.zspy")
+    file_path = zrnb.fetch_file_path(allow_download=allow_download)
     store = zarr.ZipStore(file_path)  # for loading from zip
     return hs.load(store, **kwargs)
 
@@ -133,7 +133,7 @@ def twinned_nanowire(allow_download=False, **kwargs):  # pragma: no cover
     """
     nanowire = Dataset("twinned_nanowire.hdf5")
     file_path = nanowire.fetch_file_path(allow_download=allow_download)
-    return hs.load(file_path, **kwargs)
+    return hs.load(file_path, **kwargs, reader="hspy")
 
 
 def sample_with_g(allow_download=False, **kwargs):  # pragma: no cover