+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Fixed
 - Update pyxem to work with hyperspy 2.0.0 (#969)
 - Fixed slow markers (#969)
 - Removed parallel and max_workers keywords in favor of using dask (#969)
+- Fix :meth:`~pyxem.data.zrnb_precipitate` dataset to point to proper dataset
 
 Removed
 -------